refactor(sidebar): extract path formatter and directory text out of ContentView
ContentView.swift sits near its CI-enforced file-length budget and
already contains the giant `TabItemView` plus a long tail of unrelated
sidebar helpers. `SidebarPathFormatter` and `SidebarDirectoryText` are
self-contained path-display logic with no captured state from the
surrounding view, so they belong next to the other `Sources/Sidebar/`
support files. Drops the `private` access on `SidebarDirectoryText`
(now reachable from its consumer in ContentView) and registers both
new files in the Xcode project. ContentView shrinks by 95 lines.
