+++ Sources/ContentView.swift
@@ -12163,101 +12163,6 @@ private struct SidebarEmptyArea: View {
     }
 }
 
-enum SidebarPathFormatter {
-    static let homeDirectoryPath: String = FileManager.default.homeDirectoryForCurrentUser.path
-
-    static func shortenedPath(
-        _ path: String,
-        homeDirectoryPath: String = Self.homeDirectoryPath
-    ) -> String {
-        let trimmed = path.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !trimmed.isEmpty else { return path }
-        if trimmed == homeDirectoryPath {
-            return "~"
-        }
-        if trimmed.hasPrefix(homeDirectoryPath + "/") {
-            return "~" + trimmed.dropFirst(homeDirectoryPath.count)
-        }
-        return trimmed
-    }
-
-    // The shortest single-string form. Falls back to the abbreviated path
-    // unchanged when there are no leading segments to drop, so `/tmp` stays
-    // `/tmp` rather than becoming `…/tmp`.
-    static func lastSegmentPath(
-        _ path: String,
-        homeDirectoryPath: String = Self.homeDirectoryPath
-    ) -> String {
-        pathCandidates(path, homeDirectoryPath: homeDirectoryPath).last
-            ?? shortenedPath(path, homeDirectoryPath: homeDirectoryPath)
-    }
-
-    // Ordered longest → shortest. The first entry is the full abbreviated path
-    // (with `~/` if applicable). Each subsequent entry drops one more leading
-    // segment and is prefixed with `…/`. Suitable as `ViewThatFits` candidates.
-    static func pathCandidates(
-        _ path: String,
-        homeDirectoryPath: String = Self.homeDirectoryPath
-    ) -> [String] {
-        let trimmed = path.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !trimmed.isEmpty else { return [] }
-        let abbreviated = shortenedPath(trimmed, homeDirectoryPath: homeDirectoryPath)
-        guard !abbreviated.isEmpty else { return [] }
-        if abbreviated == "~" || abbreviated == "/" { return [abbreviated] }
-
-        let prefixLength: Int = {
-            if abbreviated.hasPrefix("~/") { return 2 }
-            if abbreviated.hasPrefix("/") { return 1 }
-            return 0
-        }()
-        let suffix = String(abbreviated.dropFirst(prefixLength))
-        let parts = suffix.split(separator: "/", omittingEmptySubsequences: true).map(String.init)
-        var candidates = [abbreviated]
-        guard parts.count > 1 else { return candidates }
-        for dropCount in 1..<parts.count {
-            let remainder = parts[dropCount...].joined(separator: "/")
-            candidates.append("…/\(remainder)")
-        }
-        return candidates
-    }
-}
-
-// Picks the longest directory candidate that fits the available width.
-// Non-fallback candidates use `.fixedSize(horizontal: true)` so a candidate
-// that would only fit by truncating reports its full intrinsic width to
-// `ViewThatFits` and gets skipped in favor of the next, shorter form. The
-// final fallback keeps `.truncationMode(.tail)` for the rare case where even
-// `…/<lastSegment>` overflows.
-private struct SidebarDirectoryText: View {
-    let candidates: [String]
-    let color: Color
-
-    var body: some View {
-        if candidates.count <= 1 {
-            Text(candidates.first ?? "")
-                .font(.system(size: 10, design: .monospaced))
-                .foregroundColor(color)
-                .lineLimit(1)
-                .truncationMode(.tail)
-        } else {
-            ViewThatFits(in: .horizontal) {
-                ForEach(Array(candidates.dropLast()), id: \.self) { candidate in
-                    Text(candidate)
-                        .font(.system(size: 10, design: .monospaced))
-                        .foregroundColor(color)
-                        .lineLimit(1)
-                        .fixedSize(horizontal: true, vertical: false)
-                }
-                Text(candidates.last ?? "")
-                    .font(.system(size: 10, design: .monospaced))
-                    .foregroundColor(color)
-                    .lineLimit(1)
-                    .truncationMode(.tail)
-            }
-        }
-    }
-}
-
 enum SidebarWorkspaceShortcutHintMetrics {
     private static let measurementFont = NSFont.systemFont(ofSize: 10, weight: .semibold)
     private static let minimumSlotWidth: CGFloat = 28
