+++ Sources/Sidebar/SidebarDirectoryText.swift
@@ -1,0 +1,37 @@
+import SwiftUI
+
+// Picks the longest directory candidate that fits the available width.
+// Non-fallback candidates use `.fixedSize(horizontal: true)` so a candidate
+// that would only fit by truncating reports its full intrinsic width to
+// `ViewThatFits` and gets skipped in favor of the next, shorter form. The
+// final fallback keeps `.truncationMode(.tail)` for the rare case where even
+// `…/<lastSegment>` overflows.
+struct SidebarDirectoryText: View {
+    let candidates: [String]
+    let color: Color
+
+    var body: some View {
+        if candidates.count <= 1 {
+            Text(candidates.first ?? "")
+                .font(.system(size: 10, design: .monospaced))
+                .foregroundColor(color)
+                .lineLimit(1)
+                .truncationMode(.tail)
+        } else {
+            ViewThatFits(in: .horizontal) {
+                ForEach(Array(candidates.dropLast()), id: \.self) { candidate in
+                    Text(candidate)
+                        .font(.system(size: 10, design: .monospaced))
+                        .foregroundColor(color)
+                        .lineLimit(1)
+                        .fixedSize(horizontal: true, vertical: false)
+                }
+                Text(candidates.last ?? "")
+                    .font(.system(size: 10, design: .monospaced))
+                    .foregroundColor(color)
+                    .lineLimit(1)
+                    .truncationMode(.tail)
+            }
+        }
+    }
+}
